+++ src/Psl/Password/Exception/RuntimeException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Psl\Password\Exception;
+
+use Psl\Exception;
+
+final class RuntimeException extends Exception\RuntimeException implements ExceptionInterface
+{
+}